+++ lib/Backend/GroupBackend.php
@@ -21,12 +21,15 @@
 
 namespace OCA\UserSQL\Backend;
 
-use OC\Group\Backend;
 use OCA\UserSQL\Cache;
 use OCA\UserSQL\Constant\DB;
 use OCA\UserSQL\Model\Group;
 use OCA\UserSQL\Properties;
 use OCA\UserSQL\Repository\GroupRepository;
+use OCP\Group\Backend\ABackend;
+use OCP\Group\Backend\ICountUsersBackend;
+use OCP\Group\Backend\IGroupDetailsBackend;
+use OCP\Group\Backend\IIsAdminBackend;
 use OCP\ILogger;
 
 /**
@@ -34,7 +37,10 @@ use OCP\ILogger;
  *
  * @author Marcin Łojewski <dev@mlojewski.me>
  */
-final class GroupBackend extends Backend
+final class GroupBackend extends ABackend implements
+    ICountUsersBackend,
+    IGroupDetailsBackend,
+    IIsAdminBackend
 {
     /**
      * @var string The application name.
@@ -128,14 +134,9 @@ final class GroupBackend extends Backend
     }
 
     /**
-     * Returns the number of users in given group matching the search term.
-     *
-     * @param string $gid    The group ID.
-     * @param string $search The search term.
-     *
-     * @return int The number of users in given group matching the search term.
+     * @inheritdoc
      */
-    public function countUsersInGroup($gid, $search = "")
+    public function countUsersInGroup(string $gid, string $search = ""): int
     {
         $this->logger->debug(
             "Entering countUsersInGroup($gid, $search)",
@@ -355,17 +356,17 @@ final class GroupBackend extends Backend
     }
 
     /**
-     * Checks if a user is in the admin group.
-     *
-     * @param string $uid User ID.
-     *
-     * @return bool TRUE if a user is in the admin group, FALSE otherwise.
+     * @inheritdoc
      */
-    public function isAdmin($uid)
+    public function isAdmin(string $uid = null): bool
     {
         $this->logger->debug(
             "Entering isAdmin($uid)", ["app" => $this->appName]
         );
+
+        if (empty($this->properties[DB::GROUP_ADMIN_COLUMN]) || $uid === null) {
+            return false;
+        }
 
         $cacheKey = self::class . "admin_" . $uid;
         $admin = $this->cache->get($cacheKey);
@@ -394,17 +395,17 @@ final class GroupBackend extends Backend
     }
 
     /**
-     * Get associative array of the group details.
-     *
-     * @param string $gid The group ID.
-     *
-     * @return array Associative array of the group details.
+     * @inheritdoc
      */
-    public function getGroupDetails($gid)
+    public function getGroupDetails(string $gid): array
     {
         $this->logger->debug(
             "Entering getGroupDetails($gid)", ["app" => $this->appName]
         );
+
+        if (empty($this->properties[DB::GROUP_NAME_COLUMN])) {
+            return [];
+        }
 
         $group = $this->getGroup($gid);
 
@@ -419,21 +420,6 @@ final class GroupBackend extends Backend
         );
 
         return $details;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function getSupportedActions()
-    {
-        $actions = parent::getSupportedActions();
-
-        $actions &= empty($this->properties[DB::GROUP_ADMIN_COLUMN])
-            ? ~Backend::IS_ADMIN : ~0;
-        $actions &= empty($this->properties[DB::GROUP_NAME_COLUMN])
-            ? ~Backend::GROUP_DETAILS : ~0;
-
-        return $actions;
     }
 
     /**