+++ CHANGELOG.md
@@ -16,10 +16,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Example SQL script in README file
 - Fixed misspelling
 
+### Changed
+- Support for Nextcloud 14 only
+- Group backend implementation
+- User backend implementation
+
 ### Fixed
 - Table and column autocomplete in settings panel
 
-## [v4.0.0-rc2] - 2018-06-14
+## [4.0.0-rc2] - 2018-06-14
 ### Added
 - User active column
 
@@ -79,6 +84,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Supported version of ownCloud, Nextcloud: ownCloud 10, Nextcloud 12
 
 [Unreleased]: https://github.com/nextcloud/user_sql/compare/v4.0.0-rc2...develop
-[v4.0.0-rc2]: https://github.com/nextcloud/user_sql/compare/v4.0.0-rc1...v4.0.0-rc2
+[4.0.0-rc2]: https://github.com/nextcloud/user_sql/compare/v4.0.0-rc1...v4.0.0-rc2
 [4.0.0-rc1]: https://github.com/nextcloud/user_sql/compare/v3.1.0...v4.0.0-rc1
 [3.1.0]: https://github.com/nextcloud/user_sql/compare/v2.4.0...v3.1.0